+++ fpsim/defaults.py
@@ -99,23 +99,15 @@ person_defaults = dict(
     mothers              = -1,
 )
 
-#%% Age bins for different method switching matrices
-method_age_mapping = {
-    '<18':   [ 0, 18],
-    '18-20': [18, 20],
-    '21-25': [20, 25],
-    '>25':   [25, max_age+1], # +1 since we're using < rather than <=
-}
-
-postpartum_mapping = {
+# Postpartum keys to months
+postpartum_map = {
     'pp0to5':   [ 0, 6],
     'pp6to11':  [6, 12],
     'pp12to23': [12, 24]
 }
 
 ## Age bins for tracking age-specific fertility rate
-
-age_bin_mapping = {
+age_bin_map = {
     '10-14':   [10, 15],
     '15-19':   [15, 20],
     '20-24':   [20, 25],
@@ -126,9 +118,36 @@ age_bin_mapping = {
     '45-49':   [45, 50]
 }
 
+# Age and parity splines
 spline_ages      = np.arange(max_age + 1)
 spline_preg_ages = np.arange(max_age_preg + 1)
 spline_parities  = np.arange(max_parity + 1)
-debug_states = ["alive", "breastfeed_dur", "gestation", "lactating", "lam", "postpartum", "pregnant", "sexually_active", "postpartum_dur", \
-                "parity", "method", "age", "first_birth_age", "dobs", "age", "sexual_debut", "sexual_debut_age"]
-method_list = ["None", "Pill", "IUDs", "Injectable", "Condoms", "BTL", "Rhythm", "Withdrawal", "Implants", "Other"]
+
+# Define allowable keys to select all (all ages, all methods, etc)
+none_all_keys = [None, 'all', ':', [None], ['all'], [':']]
+
+# Definition of contraceptive methods and corresponding numbers -- can be overwritten by locations
+method_map = {
+    'None'              : 0,
+    'Pill'              : 1,
+    'IUDs'              : 2,
+    'Injectables'       : 3,
+    'Condoms'           : 4,
+    'BTL'               : 5,
+    'Withdrawal'        : 6,
+    'Implants'          : 7,
+    'Other traditional' : 8,
+    'Other modern'      : 9,
+}
+
+# Age bins for different method switching matrices -- can be overwritten by locations
+method_age_map = {
+    '<18':   [ 0, 18],
+    '18-20': [18, 20],
+    '21-25': [20, 25],
+    '>25':   [25, max_age+1], # +1 since we're using < rather than <=
+}
+
+# Finally, create default parameters to use for accessing keys etc
+default_pars = pars()
+par_keys = default_pars.keys()