+++ CHANGELOG.rst
@@ -9,6 +9,13 @@ All notable changes to the codebase are documented in this file. Changes that ma
    :depth: 1
 
 
+Version 0.11.5 (2022-05-21)
+--------------------------
+- Improvements to the scenarios, including more helpful docstrings and error messages.
+- Improved error checking of sims.
+- *GitHub info*: PR `502 <https://github.com/amath-idm/fp_analyses/pull/502>`__
+
+
 Version 0.11.4 (2022-05-20)
 --------------------------
 - Renamed parameter ``n`` to ``n_agents``, and adds parameter ``scaled_pop``.